% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{augment_bins}
\alias{augment_bins}
\title{Augment a dataframe with columns to allow plotting of conditioned box plots.}
\usage{
augment_bins(df, var, nbins = 25)
}
\arguments{
\item{df}{a dataframe}

\item{var}{a tidy-select column specification}

\item{nbins}{the number of bins to generate}
}
\value{
a tibble with two new columns, for the var's bin number and the var's
         median
}
\description{
Augment a dataframe with columns to allow plotting of conditioned box plots.
}
