% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negated value matching}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or `NULL`; the values to be matched.}

\item{table}{vector or `NULL`; the values to be matched against.}
}
\value{
a vector of the same length as x
}
\description{
`%nin%` is a negated version of `%in%`;.
}
