% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{like}
\alias{like}
\title{Convenience functions for calling grep}
\usage{
like(vector, pattern, ignore.case = FALSE, fixed = FALSE, perl = FALSE)
}
\arguments{
\item{vector}{Either a character or factor vector}

\item{pattern}{Pattern to be matched}

\item{ignore.case}{logical: is the matching to be case-insensitive?}

\item{fixed}{logical: is `pattern` to be interpreted as a literal string?}

\item{perl}{logical: is the pattern a Perl-compatible regular expression?}
}
\value{
logical vector, TRUE for items that match the pattern
}
\description{
This function is a copy of the one in data.table. It is intended to be used
through the associated operator, \%like\%.
}
