% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfptools-package.R
\docType{package}
\name{mfptools-package}
\alias{mfptools}
\alias{mfptools-package}
\title{mfptools: Marc's Analysis Tools}
\description{
A collection of tools for analysis and for creating plots.
}
\author{
\strong{Maintainer}: Marc Paterno \email{marc.paternot@gmail.com} (\href{https://orcid.org/0000-0003-0808-8388}{ORCID})

}
\keyword{internal}
