% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_benchmark}
\alias{read_benchmark}
\title{Read benchmark output generated by nanobench,}
\usage{
read_benchmark(filename)
}
\arguments{
\item{filename}{the name of the file with nanobench output to read.}
}
\value{
a tibble
}
\description{
This function expects the output in the style written for Linux machines,
which includes columns not collected on macOS.
}
